ansible: Use snapshots of the Prosody package repo
This aims to improve reproducibility of builds and remove variance due
to changes in the Prosody repo. It also improves security since the
method used to install the prosody-trunk package would not have any way
to verify any signatures.

This does require some coordination to create new snapshots, but I'm
sure the Prosody developers are nice and helpful.