mod_update_check: Fix for unbound API difference
TXT records can consist of multiple length-prefixed strings, terminated
by a zero-length string, but in most cases there is simply one.

When not using luaunbound, `record.txt` was always the first of those
strings, which should be considered a bug in dns.lua. Prosodys net.dns
translates it to an array of strings, which means this code tried to
call `match` on an array instead of a string, causing an error. These
arrays have a metatable with a `__tostring` field set to `table.concat`,
so just passing this string or array trough tostring should always
produce a sensible string.

Thanks to deimos for reporting