prosody: Switch to included mod_smacks
The XEP-0198 implementation mod_smacks was merged into prosody and
touched up a bit. It now has a limit on the amount of memory it can
consume, more efficient scheduling of ack requests and limited support
for persisting information across restarts to allow, not full
resumption, but knowing if any messages were lost.