snikket-turn-addresses: Update script to use libunbound (thanks Aerion)
We are transitioning everything to use libunbound, as it is more robust.
Indeed the previous script was failing when a DNS server was unreachable.

libunbound performs CNAME resolution automatically, so that loop has been
dropped.

Finally, a bit more error handling and input checking was sprinkled over
the result. It appears to work!