Delegate authorization from groups.$DOMAIN to $DOMAIN
This allows us to set the avatar of circles both from the app and from
the web portal (with snikket-web-portal#150).

However, this also makes all admins owners in all circles. This may be
problematic, or may not be. The upside is that it also automatically
allows managing the avatar through the apps. The downside is that it
exposes a bunch of dangerous controls (banning, kicking) which desyncs
the group membership from MUC membership.

We might want a reconciliation loop for that, or figure out something
else which overrides outcast-ness or forbids banning or stuff like that,
*or* which alternatively reflects that change in the circle UI.