prosody: Reduce non-streaming HTTP uploads from 10MB to 1MB
This improves memory usage significantly on small servers, as it will now
cause uploads over 1MB to use streaming mode or (if streaming unsupported)
just reject them.

This may have an impact on large BOSH payloads. For example, if setting a
large avatar. However such an avatar is likely too large for the network
anyway, and should be reduced by clients before upload.