Truly use Lua 5.4 for all things Prosody
The previous implementation in #137 was incomplete, because it did not
change the supervisord config.

In addition, due to luarocks being pulled in and being generally
unpleasant, depending on lua5.{1,2,3} (but notably, not on lua5.4),
causes lua 5.1 to appear on the system and for some odd reason being
chosen as the default.

Using update-alternatives fixes that.