Restore longer access token lifetime
The Prosody OAuth 2.0 module reduced the default lifetime of access
tokens to one hour, with the intent that longer lived refresh tokens
would be used to retrieve new access tokens when they expire. While the
web portal does not do this, it is somewhat annoying to have to sign in
again every hour, so restore the previous access token lifetime.