Add support for moving data to/from SQLite
Fixes #47

This allows configuring the Prosody storage module to use between the
current default of 'internal' and 'sql' with SQLite, including automatic
migration between the two. This allows those who wish to opt-in as well
as easy testing and evaluation of SQLite as future default with a chance
to go back if it does not live up to expectations.