Enable delayed deletion of user accounts deleted via IBR
After an in-band account deletion request (i.e. from an app connected to the
account) we will now disable the account, and then delete it permanently after
the configured data retention period.

During this grace period the account can be restored by a server admin if the
deletion was accidental or malicious.