prosody: Add tweak to enable mod_s2s_status
This is planned to become the basis of a federation status overview in the UI,
so admins can see which other domains their instance is federating with and
whether there are any problems.

Adding the tweak now enables us to easily test the module more widely and
ensure we have the data already available once the UI is added.