mod_invites_bootstrap: Update to new roles syntax
It is now an array of roles (the first one being the primary role). While
using the old syntax, mod_invites will store the role in the old format which
requires a migration step before it takes effect (the migration step happens
at startup, but that is likely to be before the user registers).

Thanks Zash and Root.