mod_health_report: Submit regular health reports to an API (default disabled)
This can be used to allow an external system to monitor the health of the
instance. Currently it reports the launch time, whether the server's last
exit was unclean (i.e. whether it crashed), server version (to determine
whether it needs updating), and active user counts (to gauge utilization).

Potential future data includes certificate issues or other warnings, and the
bootstrap status.